requirements.txt
streamlit
ibm-watsonx-ai
python-dotenv